+++ src/app/services/housing/housing.interface.ts
