+++ src/app/services/housing/housing.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { HousingService } from '../services/housing.service';
+import { HousingService } from './housing.service';
 
 describe('HousingService', () => {
   let service: HousingService;
