+++ src/app/services/housing/housing.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { HousingLocationInfo } from '../housinglocation.model';
+import { HousingLocationInfo } from './housing.interface';
 
 @Injectable({
   providedIn: 'root'
@@ -7,22 +7,25 @@ import { HousingLocationInfo } from '../housinglocation.model';
 export class HousingService {
   readonly baseUrl = 'https://angular.dev/assets/images/tutorials/common';
   url = 'http://localhost:3000/locations';
+  housingLocationList: HousingLocationInfo[] = [];
+
   async getAllHousingLocations(): Promise<HousingLocationInfo[]> {
     const data = await fetch(this.url);
+
     return (await data.json()) ?? []
   };
+
   async getHousingLocationById(id: number): Promise<HousingLocationInfo | undefined> {
     const data = await fetch(`${this.url}?id=${id}`);
     const locationJson = await data.json();
+
     return locationJson[0] ?? {}
   }
+
   submitApplication(firstName: string, lastName: string, email: string) {
     console.log(
       `Homes application received: firstName: ${firstName}, lastName: ${lastName}, email: ${email}.`,
     );
   }
 
-  housingLocationList: HousingLocationInfo[] = [
-   
-  ];
 }
