+++ src/app/services/housing/index.ts
@@ -1,0 +1,2 @@
+export { HousingService } from './housing.service';
+export type { HousingLocationInfo } from './housing.interface';
